+++ src/client/main.js
@@ -28,13 +28,13 @@
 import CMS from './cms';
 
 // Import CMS plugins
-import MastodonShare from './addons/mastodon_share';
 import PageBodyClass from './addons/pagebodyclass';
 import CMSSearchElement from './addons/cms-search';
 import CMSAuthorElement from './addons/cms-author';
 import CMSPagelistElement from './addons/cms-pagelist';
 import CMSButtonElement from './addons/cms-button';
 import CMSIconElement from './addons/cms-icon';
+import CMSMastodonShareElement from './addons/cms-mastodon-share';
 
 // Import specific MD renderer system
 import remarkable from './addons/loader-remarkable';
@@ -46,11 +46,11 @@ customElements.define('cms-pagelist', CMSPagelistElement);
 customElements.define('cms-search', CMSSearchElement, { extends: 'input' });
 customElements.define('cms-button', CMSButtonElement, {extends: 'a'});
 customElements.define('cms-icon', CMSIconElement, {extends: 'i'});
+customElements.define('cms-mastodon-share', CMSMastodonShareElement, {extends: 'a'});
 
 
 // Load addons
 let systemPlugins = {
-	mastodon_share: new MastodonShare(),
 	pagebodyclass: new PageBodyClass(),
 	remarkable: {
 		init: (cms) => {